requirements.txt
streamlit
opencv-python
Pillow
